+++ packages/feature-toggle/src/constants.ts
@@ -16,6 +16,7 @@ export const tokenToggleGuid: ExtensionConfigKeys = {
     'sap.ux.help.testBetaFeatures.enableFioriAIJoule': '165a0e31-35ea-4bee-8d47-b8593435a82e',
     'sap.ux.help.testBetaFeatures.enableFioriAIVsCode': '165a0e31-35ea-4bee-8d47-b8593435a82f',
     'sap.ux.help.testBetaFeatures.enableFioriAIAppModeler': '165a0e31-35ea-4bee-8d47-b8593435a82g',
+    'sap.ux.help.testBetaFeatures.enableFioriAIRapGeneration': '165a0e31-35ea-4bee-8d47-b8593435a82h',
     'sap.ux.applicationModeler.testBetaFeatures.manifestEditor': true,
     'sap.ux.appGenerator.testBetaFeatures.newAnnotationAPI': true
 } as ExtensionConfigKeys;