+++ packages/generator-adp/test/app.test.ts
@@ -3,28 +3,33 @@ import { join } from 'path';
 import { rimraf } from 'rimraf';
 import yeomanTest from 'yeoman-test';
 import { exec } from 'child_process';
+import Generator from 'yeoman-generator';
 
-import * as Logger from '@sap-ux/logger';
-import { isAppStudio } from '@sap-ux/btp-utils';
-import type { ToolsLogger } from '@sap-ux/logger';
-import type { SourceApplication, VersionDetail } from '@sap-ux/adp-tooling';
-import { getCredentialsFromStore } from '@sap-ux/system-access';
-import type { AbapServiceProvider } from '@sap-ux/axios-extension';
-import { sendTelemetry, getHostEnvironment, hostEnvironment } from '@sap-ux/fiori-generator-shared';
 import {
+    SourceManifest,
     SystemLookup,
-    getProviderConfig,
+    fetchPublicVersions,
     getConfiguredProvider,
+    getProviderConfig,
     loadApps,
     validateUI5VersionExists
 } from '@sap-ux/adp-tooling';
+import * as Logger from '@sap-ux/logger';
+import { isAppStudio } from '@sap-ux/btp-utils';
+import type { ToolsLogger } from '@sap-ux/logger';
+import type { Manifest } from '@sap-ux/project-access';
+import { getCredentialsFromStore } from '@sap-ux/system-access';
+import type { Language, SourceApplication, VersionDetail } from '@sap-ux/adp-tooling';
+import { type AbapServiceProvider, AdaptationProjectType } from '@sap-ux/axios-extension';
+import { getHostEnvironment, hostEnvironment, sendTelemetry } from '@sap-ux/fiori-generator-shared';
 
 import adpGenerator from '../src/app';
 import { initI18n, t } from '../src/utils/i18n';
+import type { JsonInput } from '../src/app/types';
 import { EventName } from '../src/telemetryEvents';
 import type { AdpGeneratorOptions } from '../src/app';
-import { getDefaultProjectName } from '../src/app/questions/helper/default-values';
 import { ConfigPrompter } from '../src/app/questions/configuration';
+import { getDefaultProjectName } from '../src/app/questions/helper/default-values';
 
 jest.mock('@sap-devx/feature-toggle-node', () => ({
     // Is BAS this will mean that the layer is CUSTOMER_BASE
@@ -36,9 +41,20 @@ jest.mock('../src/app/questions/helper/default-values.ts', () => ({
     getDefaultProjectName: jest.fn()
 }));
 
+jest.mock('../src/app/questions/helper/validators.ts', () => ({
+    ...jest.requireActual('../src/app/questions/helper/validators.ts'),
+    validateExtensibilityGenerator: jest.fn().mockReturnValue(true)
+}));
+
+jest.mock('../src/app/extension-project/index.ts', () => ({
+    ...jest.requireActual('../src/app/extension-project/index.ts'),
+    resolveNodeModuleGenerator: jest.fn().mockReturnValue('my-generator-path')
+}));
+
 jest.mock('../src/app/questions/helper/conditions', () => ({
     ...jest.requireActual('../src/app/questions/helper/conditions'),
-    showApplicationQuestion: jest.fn().mockReturnValue(true)
+    showApplicationQuestion: jest.fn().mockReturnValue(true),
+    showExtensionProjectQuestion: jest.fn().mockReturnValue(true)
 }));
 
 jest.mock('@sap-ux/system-access', () => ({
@@ -56,7 +72,8 @@ jest.mock('@sap-ux/adp-tooling', () => ({
     getConfiguredProvider: jest.fn(),
     loadApps: jest.fn(),
     getProviderConfig: jest.fn(),
-    validateUI5VersionExists: jest.fn()
+    validateUI5VersionExists: jest.fn(),
+    fetchPublicVersions: jest.fn()
 }));
 
 jest.mock('../src/utils/deps.ts', () => ({
@@ -116,12 +133,21 @@ const answers = {
     title: 'App Title',
     ui5Version: '1.134.1',
     targetFolder: testOutputDir,
-    enableTypeScript: false
+    enableTypeScript: false,
+    shouldCreateExtProject: false
+};
+
+const activeLanguages: Language[] = [{ sap: 'value', i18n: 'DE' }];
+const adaptationProjectTypes: AdaptationProjectType[] = [AdaptationProjectType.CLOUD_READY];
+const mockManifest = { 'sap.ui5': { flexEnabled: true } } as Manifest;
+const publicVersions = {
+    latest: { version: '1.134.1' } as VersionDetail,
+    '1.134.0': { version: '1.134.0' } as VersionDetail
 };
 
 const isAbapCloudMock = jest.fn();
 const getAtoInfoMock = jest.fn();
-const getSystemInfoMock = jest.fn();
+const getSystemInfoMock = jest.fn().mockResolvedValue({ adaptationProjectTypes, activeLanguages });
 const dummyProvider = {
     isAbapCloud: isAbapCloudMock,
     getAtoInfo: getAtoInfoMock,
@@ -139,10 +165,18 @@ const loggerMock: ToolsLogger = {
 } as Partial<ToolsLogger> as ToolsLogger;
 jest.spyOn(Logger, 'ToolsLogger').mockImplementation(() => loggerMock);
 
+const executeCommandSpy = jest.fn();
+const vscodeMock = {
+    commands: {
+        executeCommand: executeCommandSpy
+    }
+};
+
 const loadAppsMock = loadApps as jest.Mock;
 const execMock = exec as unknown as jest.Mock;
 const mockIsAppStudio = isAppStudio as jest.Mock;
 const getProviderConfigMock = getProviderConfig as jest.Mock;
+const fetchPublicVersionsMock = fetchPublicVersions as jest.Mock;
 const sendTelemetryMock = sendTelemetry as jest.Mock;
 const getHostEnvironmentMock = getHostEnvironment as jest.Mock;
 const getDefaultProjectNameMock = getDefaultProjectName as jest.Mock;
@@ -159,13 +193,12 @@ describe('Adaptation Project Generator Integration Test', () => {
         loadAppsMock.mockResolvedValue(apps);
         jest.spyOn(ConfigPrompter.prototype, 'provider', 'get').mockReturnValue(dummyProvider);
         jest.spyOn(ConfigPrompter.prototype, 'ui5', 'get').mockReturnValue({
-            publicVersions: {
-                latest: { version: '1.134.1' } as VersionDetail,
-                '1.134.0': { version: '1.134.0' } as VersionDetail
-            },
+            publicVersions,
             ui5Versions: ['1.134.1 (latest)', '1.134.0'],
-            systemVersion: '1.136.0.204546979753'
+            systemVersion: '1.136.0'
         });
+        jest.spyOn(ConfigPrompter.prototype, 'manifest', 'get').mockReturnValue(mockManifest);
+        jest.spyOn(SourceManifest.prototype, 'getManifest').mockResolvedValue(mockManifest);
         validateUI5VersionExistsMock.mockReturnValue(true);
         jest.spyOn(SystemLookup.prototype, 'getSystems').mockResolvedValue(endpoints);
         jest.spyOn(SystemLookup.prototype, 'getSystemRequiresAuth').mockResolvedValue(false);
@@ -180,6 +213,8 @@ describe('Adaptation Project Generator Integration Test', () => {
 
         getDefaultProjectNameMock.mockReturnValue('app.variant1');
         getCredentialsFromStoreMock.mockResolvedValue(undefined);
+
+        fetchPublicVersionsMock.mockResolvedValue(publicVersions);
     });
 
     beforeAll(async () => {
@@ -208,19 +243,43 @@ describe('Adaptation Project Generator Integration Test', () => {
         await expect(runContext.run()).rejects.toThrow(t('error.updatingApp'));
     });
 
+    it('should call composeWith to generate an extension project in case the application is not supported', async () => {
+        mockIsAppStudio.mockReturnValue(false);
+        const composeWithSpy = jest.spyOn(Generator.prototype, 'composeWith').mockReturnValue([]);
+
+        const runContext = yeomanTest
+            .create(adpGenerator, { resolved: generatorPath }, { cwd: testOutputDir })
+            .withOptions({ shouldInstallDeps: false, vscode: vscodeMock } as AdpGeneratorOptions)
+            .withPrompts({ ...answers, shouldCreateExtProject: true });
+
+        await expect(runContext.run()).resolves.not.toThrow();
+
+        expect(composeWithSpy).toHaveBeenCalledWith(
+            expect.stringMatching(''),
+            expect.objectContaining({
+                arguments: [
+                    '{"destination":{"name":"SystemA"},"applicationNS":"customer.app.variant","applicationName":"app.variant","userUI5Ver":"1.134.1","namespace":"sap.ui.demoapps.f1"}'
+                ],
+                appWizard: expect.anything()
+            })
+        );
+    });
+
     it('should generate an onPremise adaptation project successfully', async () => {
         mockIsAppStudio.mockReturnValue(false);
 
         const runContext = yeomanTest
             .create(adpGenerator, { resolved: generatorPath }, { cwd: testOutputDir })
-            .withOptions({ shouldInstallDeps: true } as AdpGeneratorOptions)
+            .withOptions({ shouldInstallDeps: true, vscode: vscodeMock } as AdpGeneratorOptions)
             .withPrompts(answers);
 
         await expect(runContext.run()).resolves.not.toThrow();
 
+        expect(executeCommandSpy).toHaveBeenCalledTimes(1);
+
         const generatedDirs = fs.readdirSync(testOutputDir);
-        expect(generatedDirs.length).toBeGreaterThan(0);
-        const projectFolder = join(testOutputDir, generatedDirs[0]);
+        expect(generatedDirs).toContain(answers.projectName);
+        const projectFolder = join(testOutputDir, answers.projectName);
 
         const manifestPath = join(projectFolder, 'webapp', 'manifest.appdescr_variant');
         const i18nPath = join(projectFolder, 'webapp', 'i18n', 'i18n.properties');
@@ -245,5 +304,45 @@ describe('Adaptation Project Generator Integration Test', () => {
             }),
             projectFolder
         );
+    });
+
+    it('should create adaptation project from json correctly', async () => {
+        const jsonInput: JsonInput = {
+            system: 'urlA',
+            username: 'user1',
+            password: 'pass1',
+            client: '010',
+            application: 'sap.ui.demoapps.f1',
+            projectName: 'my.app',
+            namespace: 'customer.my.app',
+            applicationTitle: 'My app title',
+            targetFolder: testOutputDir
+        };
+        const jsonInputString = JSON.stringify(jsonInput);
+
+        const runContext = yeomanTest
+            .create(adpGenerator, { resolved: generatorPath }, { cwd: testOutputDir })
+            .withArguments([jsonInputString]);
+
+        await expect(runContext.run()).resolves.not.toThrow();
+
+        const generatedDirs = fs.readdirSync(testOutputDir);
+        expect(generatedDirs).toContain(jsonInput.projectName);
+        const projectFolder = join(testOutputDir, jsonInput.projectName!);
+
+        const manifestPath = join(projectFolder, 'webapp', 'manifest.appdescr_variant');
+        const i18nPath = join(projectFolder, 'webapp', 'i18n', 'i18n.properties');
+        const ui5Yaml = join(projectFolder, 'ui5.yaml');
+
+        expect(fs.existsSync(manifestPath)).toBe(true);
+        expect(fs.existsSync(i18nPath)).toBe(true);
+        expect(fs.existsSync(ui5Yaml)).toBe(true);
+
+        const manifestContent = fs.readFileSync(manifestPath, 'utf8');
+        const i18nContent = fs.readFileSync(i18nPath, 'utf8');
+        const ui5Content = fs.readFileSync(ui5Yaml, 'utf8');
+        expect(manifestContent).toMatchSnapshot();
+        expect(i18nContent).toMatchSnapshot();
+        expect(ui5Content).toMatchSnapshot();
     });
 });