+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/odata-cli
 
+## 0.17.0
+
+### Minor Changes
+
+-   a28357d: chore - drop node18 support as it is out of maintenance
+
+### Patch Changes
+
+-   Updated dependencies [a28357d]
+    -   @sap-ux/axios-extension@1.21.0
+    -   @sap-ux/btp-utils@1.1.0
+    -   @sap-ux/logger@0.7.0
+
+## 0.16.14
+
+### Patch Changes
+
+-   @sap-ux/axios-extension@1.20.3
+
 ## 0.16.13
 
 ### Patch Changes