+++ packages/cds-odata-annotation-converter/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sap-ux/cds-odata-annotation-converter
 
+## 0.5.8
+
+### Patch Changes
+
+-   Updated dependencies [a28357d]
+    -   @sap-ux/text-document-utils@0.3.0
+    -   @sap-ux/cds-annotation-parser@0.2.10
+    -   @sap-ux/odata-annotation-core@0.2.4
+    -   @sap-ux/odata-vocabularies@0.4.8
+
 ## 0.5.7
 
 ### Patch Changes