+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.2.0
+
+### Minor Changes
+
+-   a28357d: chore - drop node18 support as it is out of maintenance
+
+### Patch Changes
+
+-   Updated dependencies [a28357d]
+    -   @sap-ux/deploy-config-generator-shared@0.1.0
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.0
+    -   @sap-ux/cf-deploy-config-writer@0.3.0
+    -   @sap-ux/fiori-generator-shared@0.12.0
+    -   @sap-ux/inquirer-common@0.7.0
+    -   @sap-ux/feature-toggle@0.3.0
+    -   @sap-ux/project-access@1.30.0
+    -   @sap-ux/btp-utils@1.1.0
+
+## 0.1.60
+
+### Patch Changes
+
+-   Updated dependencies [5585f0d]
+    -   @sap-ux/feature-toggle@0.2.4
+    -   @sap-ux/inquirer-common@0.6.43
+    -   @sap-ux/deploy-config-generator-shared@0.0.47
+    -   @sap-ux/cf-deploy-config-inquirer@0.2.37
+
 ## 0.1.59
 
 ### Patch Changes